requirements.txt
fastapi
uvicorn[standard]
mistralai
httpx
python-dotenv
websockets
pymavlink
